add scenarios for users trying to reconfirm email changes
- add Scenario: Logged in user reconfirms an email change
- ad Scenario: Logged out user reconfirms an email change

+++ app/models/user.rb
@@ -10,11 +10,11 @@ class User < ActiveRecord::Base
 
   # Setup accessible (or protected) attributes for your model
   attr_accessible :email, :password, :password_confirmation, :remember_me, :employee_id, :confirmed_at,
-    :first_name, :last_name, :organization_name, :on_signup
+    :first_name, :last_name, :organization_name, :on_signup, :email_change_attributes, :confirming_email_change
   attr_reader :current_employee
   # Virtual attributes for registration purposes. On registration the auto-created organization
   # and employee get useful values.
-  attr_accessor :first_name, :last_name, :organization_name, :on_signup
+  attr_accessor :first_name, :last_name, :organization_name, :on_signup, :confirming_email_change
 
   validates_presence_of :first_name, :last_name, :organization_name, if: Proc.new { |u| u.on_signup }
 
@@ -22,6 +22,8 @@ class User < ActiveRecord::Base
   has_many :invitations
   has_many :organizations, :through => :employees
   has_one  :email_change
+
+  accepts_nested_attributes_for :email_change
 
   def label
     email
@@ -43,12 +45,19 @@ class User < ActiveRecord::Base
     employees.count > 1
   end
 
+  def confirming_email_change?
+    confirming_email_change
+  end
+
   def create_email_change
+    # If the user is just confirming his/her email change we do not want to create a new email change
+    return true if confirming_email_change?
     if email_was.present? and email_changed?
       email_change.destroy if email_change
       create_email_change!(email: email)
       email = email_was
     end
+    true
   end
   before_save :create_email_change
 end
